Rename: 알림메시지 컴포넌트 파일 이름 변경 및  MainChattingList.tsx 파일 위치 이동
- NotificationMessage.tsx -> NotificationCard.tsx 파일명 변경
- MainChattingList.tsx 파일 위치 components/contents 폴더 안으로 이동
