+++ src/components/contents/MainChattingList.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
-import useFetchData from "../hooks/useFetchData";
-import ChatUserBadge from "./common/ChatUserBadge";
-import EnterChatRoom from "./modal/EnterChatRoom";
-import { Post } from "../mocks/mockType";
+import useFetchData from "../../hooks/useFetchData";
+import ChatUserBadge from "../common/ChatUserBadge";
+import EnterChatRoom from "../modal/EnterChatRoom";
+import { Post } from "../../mocks/mockType";
 
 const MainChattingList = (): JSX.Element => {
   const [showModal, setShowModal] = useState(false);
