+++ src/views/NotificationPage.tsx
@@ -1,4 +1,4 @@
-import NotificationMessage from "../components/NotificationMessage";
+import NotificationCard from "../components/NotificationCard";
 
 const NotificationPage = (): JSX.Element => {
   return (
@@ -7,7 +7,7 @@ const NotificationPage = (): JSX.Element => {
         <h1 className="text-2xl font-semibold">채팅방 개설 알림</h1>
         <button>전체 삭제</button>
       </div>
-      <NotificationMessage />
+      <NotificationCard />
     </>
   );
 };
